Arrow Key Movement
Hello everyone! I've been playing around with some javascript/jquery code and thought this could be really useful for the idea we have! I'm still working out how to get the box to move like it should, right now it's sort of shaking in a corner, I'm thinking a while loop of some sort would be best case scenario. The html is very messy with inline styling because I was focusing more on functionality, I do not intend to keep it like that. Let me know if anyone figures out a better way or doesn't understand something. Have a good day!