Animations work
I've got the little guy scuttling about the screen with arrow keys, animated the stabbing and have a lot of comments and refactoring in there this time so it should be easier to understand. Can't really do more without us talking about the server and health bar together but we should be able to.